Harden some reflection invoke tests
