+++ src/libraries/System.Runtime/tests/System.Reflection.Tests/InvokeInterpreted/System.Reflection.InvokeInterpreted.Tests.csproj
@@ -14,7 +14,6 @@
     <Compile Include="..\MethodInfoTests.cs" />
     <Compile Include="..\MethodInvokerTests.cs" />
     <Compile Include="..\PropertyInfoTests.cs" />
-    <Compile Include="$(CommonTestPath)\System\Reflection\InvokeInterpretedTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\TestAssembly\TestAssembly.csproj" />
